All Done

+++ src/pages/item_page.js
@@ -1,31 +1,52 @@
 import { deleteDoc, doc, getDoc } from "firebase/firestore";
-import { useEffect, useState } from "react";
+import { useContext, useEffect, useState } from "react";
+import ReactPlayer from "react-player";
 import { Link, useNavigate, useParams } from "react-router-dom";
+import { AuthContext } from "../context/AuthContext";
 import { db } from "../firebase";
 
 export const ItemPage = () => {
 
     const { id } = useParams();
 
+    const { currentUser } = useContext(AuthContext);
+
+
     const navigate = useNavigate();
 
     const [item, setItem] = useState(null);
+    const [restaurant, setRestaurant] = useState(null);
 
     const itemsRef = doc(db, "items", id);
     const getItem = async () => {
         const data = await getDoc(itemsRef);
         if (data.exists()) {
             setItem(data.data());
+            const userRef = doc(db, "users", data.data().uid);
+            const user = await getDoc(userRef);
+            if (user.exists()) {
+                setRestaurant(user.data().restaurantName);
+            }
+
         }
         else {
             console.log("NO");
         }
     };
 
+
+
     const deleteItem = async () => {
-        await deleteDoc(doc(db, "items", id));
-        console.log("Del");
-        navigate('/menus');
+
+        if (currentUser) {
+            await deleteDoc(doc(db, "items", id));
+            console.log("Del");
+            navigate('/menus');
+
+        }
+        else{
+            navigate('/');
+        }
     };
 
     useEffect(() => {
@@ -35,7 +56,7 @@ export const ItemPage = () => {
     return (
         <div className="itemPage">
             <div className="name">
-                <h1>Restaurant Name</h1>
+                <h1>{restaurant}</h1>
             </div>
             <div className="back">
                 <button> <Link to="/menus"> Back </Link></button>
@@ -47,12 +68,20 @@ export const ItemPage = () => {
                 </div>
                 <div className="allDetails">
                     <h3> {item?.dishName} </h3>
+                    <div>
+                        <ReactPlayer
+                            className="react-player"
+                            url={item?.videoLink}
+                            width="60ex"
+                            height="40ex"
+                        />
+                    </div>
                     <p>
                         {item?.description}
                     </p>
-                    <span><b> Veg type</b>: Veg</span>
+                    <span><b> Veg type</b>: {item?.foodType}</span>
                     <span><b> Meal type</b>: {item?.mealType} </span>
-                    <span id="price"><b> Price</b>: 10 &#8377; </span>
+                    <span id="price"><b> Price</b>: {item?.price} &#8377; </span>
                     <span><b> Liked</b>: 100 </span>
                     <div>
                         <button className="edit"><Link to={`/UpdateItem/${id}`}>Edit</Link></button>
